+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/common/domain/StandardIconProvider.kt
@@ -1,7 +1,10 @@
 package io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain
 
+import androidx.compose.ui.text.font.FontWeight
 import io.github.composegears.valkyrie.ui.screen.webimport.common.model.FontByteArray
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.IconStyle
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.SizeSettings
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.StandardIcon
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.StandardIconConfig
 
 interface StandardIconProvider {
@@ -11,8 +14,9 @@ interface StandardIconProvider {
     val persistentSize: Int
 
     fun updatePersistentSize(value: Int)
+    fun resolveFontWeight(style: IconStyle?): FontWeight = FontWeight.W400
 
     suspend fun loadConfig(): StandardIconConfig
-    suspend fun loadFontBytes(): FontByteArray
-    suspend fun downloadSvg(iconName: String, settings: SizeSettings): String
+    suspend fun loadFontBytes(style: IconStyle? = null): FontByteArray
+    suspend fun downloadSvg(icon: StandardIcon, settings: SizeSettings): String
 }