+++ tools/idea-plugin/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- [Web Import] Add `Font Awesome` icons provider
+
 ### Changed
 
 - Migrate to IDEA code highlighting for tooltip snippets